RT1-9640: separate lane departure checker library
