+++ planning/autoware_departure_checker/src/departure_checker.cpp
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/lane_departure_checker/lane_departure_checker.hpp"
+#include "autoware/departure_checker/departure_checker.hpp"
 
-#include "autoware/lane_departure_checker/utils.hpp"
+#include "autoware/departure_checker/utils.hpp"
 
 #include <autoware_utils/math/normalization.hpp>
 #include <autoware_utils/math/unit_conversion.hpp>
