+++ planning/autoware_departure_checker/test/test_resample_trajectory.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/lane_departure_checker/utils.hpp"
+#include "autoware/departure_checker/utils.hpp"
 
 #include <Eigen/Core>
