+++ planning/behavior_path_planner/autoware_behavior_path_planner/package.xml
@@ -43,7 +43,6 @@
   <depend>autoware_internal_debug_msgs</depend>
   <depend>autoware_internal_planning_msgs</depend>
   <depend>autoware_interpolation</depend>
-  <depend>autoware_lane_departure_checker</depend>
   <depend>autoware_lanelet2_extension</depend>
   <depend>autoware_motion_utils</depend>
   <depend>autoware_object_recognition_utils</depend>
